Moved CompactProtocolWriter to its own file
