+++ cpp/src/io/parquet/parquet.cpp
@@ -366,100 +366,6 @@ int CompactProtocolReader::WalkSchema(
   }
 }
 
-/**
- * @Brief Parquet CompactProtocolWriter class
- */
-
-size_t CompactProtocolWriter::write(const FileMetaData &f)
-{
-  field_int(1, f.version);
-  field_struct_list(2, f.schema);
-  field_int(3, f.num_rows);
-  field_struct_list(4, f.row_groups);
-  if (f.key_value_metadata.size() != 0) { field_struct_list(5, f.key_value_metadata); }
-  if (f.created_by.size() != 0) { field_string(6, f.created_by); }
-  if (f.column_order_listsize != 0) {
-    // Dummy list of struct containing an empty field1 struct
-    put_fldh(7, current_field(), ST_FLD_LIST);
-    putb((uint8_t)((std::min(f.column_order_listsize, 0xfu) << 4) | ST_FLD_STRUCT));
-    if (f.column_order_listsize >= 0xf) put_uint(f.column_order_listsize);
-    for (uint32_t i = 0; i < f.column_order_listsize; i++) {
-      put_fldh(1, 0, ST_FLD_STRUCT);
-      putb(0);  // ColumnOrder.field1 struct end
-      putb(0);  // ColumnOrder struct end
-    }
-    set_current_field(7);
-  }
-  return value();
-}
-
-size_t CompactProtocolWriter::write(const SchemaElement &s)
-{
-  if (s.type != UNDEFINED_TYPE) {
-    field_int(1, s.type);
-    if (s.type_length != 0) { field_int(2, s.type_length); }
-  }
-  if (s.repetition_type != NO_REPETITION_TYPE) { field_int(3, s.repetition_type); }
-  field_string(4, s.name);
-
-  if (s.type == UNDEFINED_TYPE) { field_int(5, s.num_children); }
-  if (s.converted_type != UNKNOWN) {
-    field_int(6, s.converted_type);
-    if (s.converted_type == DECIMAL) {
-      field_int(7, s.decimal_scale);
-      field_int(8, s.decimal_precision);
-    }
-  }
-  return value();
-}
-
-size_t CompactProtocolWriter::write(const RowGroup &r)
-{
-  field_struct_list(1, r.columns);
-  field_int(2, r.total_byte_size);
-  field_int(3, r.num_rows);
-  return value();
-}
-
-size_t CompactProtocolWriter::write(const KeyValue &k)
-{
-  field_string(1, k.key);
-  if (k.value.size() != 0) { field_string(2, k.value); }
-  return value();
-}
-
-size_t CompactProtocolWriter::write(const ColumnChunk &s)
-{
-  if (s.file_path.size() != 0) { field_string(1, s.file_path); }
-  field_int(2, s.file_offset);
-  field_struct(3, s.meta_data);
-  if (s.offset_index_length != 0) {
-    field_int(4, s.offset_index_offset);
-    field_int(5, s.offset_index_length);
-  }
-  if (s.column_index_length != 0) {
-    field_int(6, s.column_index_offset);
-    field_int(7, s.column_index_length);
-  }
-  return value();
-}
-
-size_t CompactProtocolWriter::write(const ColumnChunkMetaData &s)
-{
-  field_int(1, s.type);
-  field_int_list(2, s.encodings);
-  field_string_list(3, s.path_in_schema);
-  field_int(4, s.codec);
-  field_int(5, s.num_values);
-  field_int(6, s.total_uncompressed_size);
-  field_int(7, s.total_compressed_size);
-  field_int(9, s.data_page_offset);
-  if (s.index_page_offset != 0) { field_int(10, s.index_page_offset); }
-  if (s.dictionary_page_offset != 0) { field_int(11, s.dictionary_page_offset); }
-  if (s.statistics_blob.size() != 0) { field_struct_blob(12, s.statistics_blob); }
-  return value();
-}
-
 }  // namespace parquet
 }  // namespace io
 }  // namespace cudf
